moved enum to GraphQL schemas,fixed post import structue and mock user inserts
